[Feature][MRV2] Support DFlash
