+++ vllm_ascend/ops/rotary_embedding.py
@@ -21,6 +21,7 @@ import os
 import torch
 import torch_npu
 from vllm.config import get_current_vllm_config
+from vllm.forward_context import is_forward_context_available
 from vllm.model_executor.layers.rotary_embedding import (
     DeepseekScalingRotaryEmbedding,
     MRotaryEmbedding,
@@ -241,8 +242,12 @@ class AscendRotaryEmbedding(RotaryEmbedding):
         is_neox_style = self.is_neox_style
         if is_neox_style_override is not None:
             is_neox_style = is_neox_style_override
-        is_draft_model = _EXTRA_CTX.is_draft_model
-        flash_comm_v1_enabled = _EXTRA_CTX.flash_comm_v1_enabled
+        is_draft_model = (
+            _EXTRA_CTX.is_draft_model if is_forward_context_available() else False
+        )
+        flash_comm_v1_enabled = (
+            _EXTRA_CTX.flash_comm_v1_enabled if is_forward_context_available() else False
+        )
         if is_draft_model and self.use_mtp and flash_comm_v1_enabled:
             positions = torch.ops.vllm.maybe_all_gather_and_maybe_unpad(positions.contiguous(), True)
         return torch.ops.vllm.npu_rotary_embedding(
