+++ vllm_ascend/worker/v2/spec_decode/__init__.py
@@ -1,0 +1,42 @@
+# Adapt from https://github.com/vllm-project/vllm/blob/main/vllm/v1/worker/gpu/sample/spec_decode/__init__.py
+# SPDX-License-Identifier: Apache-2.0
+# SPDX-FileCopyrightText: Copyright contributors to the vLLM project
+# Copyright (c) 2025 Huawei Technologies Co., Ltd. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+# This file is a part of the vllm-ascend project.
+#
+import torch
+from vllm.config import VllmConfig
+
+
+def init_speculator(
+    vllm_config: VllmConfig,
+    device: torch.device,
+):
+    """Override GPU init_speculator for Ascend NPUs.
+    Use AscendEagleSpeculator when eagle is used.
+    """
+    speculative_config = vllm_config.speculative_config
+    assert speculative_config is not None
+    if speculative_config.use_dflash():
+        from vllm.v1.worker.gpu.spec_decode.dflash.speculator import (
+            DFlashSpeculator,
+        )
+
+        return DFlashSpeculator(vllm_config, device)
+    if speculative_config.use_eagle():
+        from vllm_ascend.worker.v2.spec_decode.eagle.speculator import AscendEagleSpeculator
+
+        return AscendEagleSpeculator(vllm_config, device)
+    raise NotImplementedError(f"{speculative_config.method} is not supported yet.")
